Remove table page and move functions to utils
