+++ projects/novo-e2e/src/e2e/common/table.common.ts
@@ -8,7 +8,7 @@ import {
   sortTableColumn,
   verifyColumnFiltered,
   verifyColumnSorted,
-} from './table.common.page';
+} from 'utils/TableUtil';
 
 export async function testElements(columns: string[]) {
   for (const column of columns) {
