+++ projects/novo-e2e/src/utils/CurrencyUtil.ts
@@ -1,0 +1,7 @@
+export async function negativeCurrencyToNegativeNumber(filteredArray): Promise<Array<any>> {
+  const isNegativeCurrency = (item) => /^[(]+[\d|.|,]*[)]+$/g.test(item);
+  const convertNegativeCurrencyToNegativeNumber = (item) => `-${item.replace(/[(),]/g, '')}`;
+  return filteredArray
+    ? filteredArray.map((item) => (isNegativeCurrency(item) ? convertNegativeCurrencyToNegativeNumber(item) : item))
+    : [];
+}
